backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "authlib>=1.4.1",
    "bcrypt>=4.2.1",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.8.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
